fix:fixing user start time issue on the calendar
