+++ src/hooks/useFetchEvents.ts
@@ -1,11 +1,19 @@
 import { useState, useEffect, useCallback } from "react";
 import { auth, db } from "../../firebase";
-import { collection, query, getDocs, Timestamp } from "firebase/firestore";
+import {
+  collection,
+  query,
+  getDocs,
+  Timestamp,
+  doc,
+  getDoc,
+} from "firebase/firestore";
 import { EventInput } from "../interfaces/types";
 import { fetchFirestoreEvents } from "@/lib/converters/events";
 
 const useFetchEvents = () => {
   const [events, setEvents] = useState<EventInput[]>([]);
+  const [userStartTime, setUserStartTime] = useState("07:00:00");
   const [loading, setLoading] = useState(true);
 
   const fetchEvents = useCallback(async () => {
@@ -19,10 +27,23 @@ const useFetchEvents = () => {
     setLoading(false);
   }, []);
 
+  const fetchUserStartTime = useCallback(async () => {
+    const user = auth.currentUser;
+    if (user) {
+      const userDocRef = doc(db, "users", user.uid);
+      const userDoc = await getDoc(userDocRef);
+      if (userDoc.exists()) {
+        const userData = userDoc.data();
+        setUserStartTime(userData.calendarStartTime || "07:00:00");
+      }
+    }
+  }, []);
+
   useEffect(() => {
     const unsubscribe = auth.onAuthStateChanged((user) => {
       if (user) {
         fetchEvents();
+        fetchUserStartTime();
       } else {
         console.warn("No authenticated user found.");
         setLoading(false);
@@ -31,9 +52,9 @@ const useFetchEvents = () => {
 
     // Cleanup the subscription on component unmount
     return () => unsubscribe();
-  }, [fetchEvents]);
+  }, [fetchEvents, fetchUserStartTime]);
 
-  return { events, loading, fetchEvents };
+  return { events, userStartTime, loading, fetchEvents, fetchUserStartTime };
 };
 
 export default useFetchEvents;
